--- a/pkg/ring/ring.go
+++ b/pkg/ring/ring.go
@@ -77,33 +77,33 @@ type ReadRing interface {
 
 // Operation describes which instances can be included in the replica set, based on their state.
 type Operation interface {
-	// ShouldExtendReplicaSet returns true if given a state of instance that's going to be
+	// ShouldExtendReplicaSetOnState returns true if given a state of instance that's going to be
 	// added to the replica set, the replica set size should be extended by 1
 	// more instance for the given operation.
 	ShouldExtendReplicaSetOnState(s IngesterState) bool
 
-	// Used during "filtering" phase to remove undesired instances based on their state.
-	IncludeInstanceInState(s IngesterState) bool
+	// IsInstanceInStateHealthy is used during "filtering" phase to remove undesired instances based on their state.
+	IsInstanceInStateHealthy(s IngesterState) bool
 }
 
 var (
 	// Write operation that also extends replica set, if ingester state is not ACTIVE.
-	Write Operation = NewOpWithReplicaSetExtension(func(s IngesterState) bool {
+	Write Operation = NewOp([]IngesterState{ACTIVE}, func(s IngesterState) bool {
 		// We do not want to Write to Ingesters that are not ACTIVE, but we do want
 		// to write the extra replica somewhere.  So we increase the size of the set
-		// of replicas for the key. This means we have to also increase the
-		// size of the replica set for read, but we can read from Leaving ingesters,
-		// so don't skip it in this case.
+		// of replicas for the key.
 		// NB dead ingester will be filtered later by defaultReplicationStrategy.Filter().
 		return s != ACTIVE
-	}, ACTIVE)
+	})
 
 	// WriteNoExtend is like Write, but with no replicaset extension.
-	WriteNoExtend Operation = NewOp(ACTIVE)
+	WriteNoExtend Operation = NewOp([]IngesterState{ACTIVE}, nil)
 
-	Read Operation = NewOpWithReplicaSetExtension(func(s IngesterState) bool {
+	Read Operation = NewOp([]IngesterState{ACTIVE, PENDING, LEAVING}, func(s IngesterState) bool {
+		// To match Write with extended replica set we have to also increase the
+		// size of the replica set for Read, but we can read from LEAVING ingesters.
 		return s != ACTIVE && s != LEAVING
-	}, ACTIVE, LEAVING, PENDING)
+	})
 
 	// Reporting is a special value for inquiring about health.
 	Reporting Operation = allStatesRingOperation{}
@@ -818,26 +818,24 @@ func (r *Ring) setCachedShuffledSubring(identifier string, size int, subring *Ri
 
 // Default implementation for Operation interface.
 type Op struct {
-	includedStates map[IngesterState]bool
-	extend         func(s IngesterState) bool
+	// Bitmap of instance states that should be considered for this operation.
+	states uint32
+
+	// Function that extends replica set based on instance state.
+	extend func(s IngesterState) bool
 }
 
-func NewOp(states ...IngesterState) Op {
-	is := make(map[IngesterState]bool, len(states))
+// NewOp constructs new Operation with given "healthy" states for operation, and optional function to extend replica set.
+func NewOp(states []IngesterState, shouldExtendReplicaSet func(s IngesterState) bool) Op {
+	st := uint32(0)
 	for _, s := range states {
-		is[s] = true
+		st |= (1 << s)
 	}
-	return Op{includedStates: is, extend: nil}
+	return Op{states: st, extend: shouldExtendReplicaSet}
 }
 
-func NewOpWithReplicaSetExtension(shouldExtendReplicaSet func(s IngesterState) bool, states ...IngesterState) Op {
-	op := NewOp(states...)
-	op.extend = shouldExtendReplicaSet
-	return op
-}
-
-func (op Op) IncludeInstanceInState(s IngesterState) bool {
-	return op.includedStates[s]
+func (op Op) IsInstanceInStateHealthy(s IngesterState) bool {
+	return op.states&(1<<s) > 0
 }
 
 func (op Op) ShouldExtendReplicaSetOnState(s IngesterState) bool {
@@ -849,5 +847,5 @@ func (op Op) ShouldExtendReplicaSetOnState(s IngesterState) bool {
 
 type allStatesRingOperation struct{}
 
-func (op allStatesRingOperation) IncludeInstanceInState(_ IngesterState) bool        { return true }
+func (op allStatesRingOperation) IsInstanceInStateHealthy(_ IngesterState) bool      { return true }
 func (op allStatesRingOperation) ShouldExtendReplicaSetOnState(_ IngesterState) bool { return false }