… pkg/ring/model_test.go
… pkg/ring/replication_strategy_test.go
… pkg/ring/ring_test.go
… pkg/ring/util_test.go
… pkg/storegateway/gateway.go
… pkg/storegateway/replication_strategy.go
… pkg/storegateway/replication_strategy_test.go
… pkg/storegateway/sharding_strategy.go