**What this PR does**: This PR refactors `ring.Operation` ~to interface~, which allows us to define new ring "operations" in packages where they are used. "Extension of replica set" logic has been moved from "replication strategy" to operation itself. "Replication strategy" is now only filtering unwanted instances.

Inspired by another operation being added in #3664.

**Checklist**
- [x] Tests updated
- [na] Documentation added
- [na] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`