@pstibrany 2021-01-11
> Could you compare `ring.Get()` benchmark, please?

Using `BenchmarkRing_Get`, I see this:

```
name        old time/op    new time/op    delta
Ring_Get-4     979ns ± 5%     992ns ± 4%  +1.34%  (p=0.035 n=25+23)
```

i.e. tiny slowdown. With some changes to `Op` implementation, I can get it to:
```
name        old time/op    new time/op    delta
Ring_Get-4     979ns ± 5%     972ns ± 3%   ~     (p=0.354 n=25+22)
```