@pracucci 2021-01-11
> By removing the interface, I can get it down to:
> ...
> ie. 11% speed improvement compared to master.

If we don't need the interface, I would do it. But why is faster than `master`?