@pstibrany 2021-01-11
> If we don't need the interface, I would do it. But why is faster than `master`?

I've pushed the [change](https://github.com/cortexproject/cortex/pull/3675/commits/6ef474a6f9eb2d6943fad1e0d28ce8cac0739eaf). I haven't done any analysis except running benchmark, but I think current implementation removes some interface function calls (replication strategy for extending replica set), and switch and if-condition compared to master. Furthermore, by not using the `Operation` interface anymore, there are no interface-conversions needed.